+++ src/main/java/com/gangbean/stockservice/dto/AccountOpenRequest.java
@@ -4,15 +4,17 @@ import com.gangbean.stockservice.domain.Account;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Getter
 @NoArgsConstructor
 public class AccountOpenRequest {
     private String bankName;
     private Long bankNumber;
-    private Long balance;
+    private BigDecimal balance;
     private Long memberId;
 
-    public AccountOpenRequest(String bankName, Long bankNumber, Long balance, Long memberId) {
+    public AccountOpenRequest(String bankName, Long bankNumber, BigDecimal balance, Long memberId) {
         this.bankName = bankName;
         this.bankNumber = bankNumber;
         this.balance = balance;
@@ -31,7 +33,7 @@ public class AccountOpenRequest {
         return bankNumber;
     }
 
-    public Long balance() {
+    public BigDecimal balance() {
         return balance;
     }
 }
