+++ src/main/java/com/gangbean/stockservice/dto/StockBuyRequest.java
@@ -4,12 +4,14 @@ import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Getter
 @AllArgsConstructor
 @NoArgsConstructor
 public class StockBuyRequest {
 
-    private Long amount;
+    private BigDecimal amount;
 
-    private Long price;
+    private BigDecimal price;
 }
